bytefix_project/uni-club/backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.1
email-validator==2.2.0
watchfiles==0.21.0